Remove reference to performance cop
Performance cops were moved out in to their own gem in Rubocop 0.68[1], which
means that referring to them logs a warning like:

`Warning: unrecognized cop Performance/Casecmp found in …`

Remove the reference to the performance cop to resolve this. Note that it may
negatively impact users who do not upgrade to 0.68, as this previously disabled
Performace cop will now be enabled until they upgrade to 0.68.

[1]: https://github.com/rubocop-hq/rubocop/issues/5977