Add C5 conventions for configuring rubocop-rspec (#15)
In general the out-of-the-box rules lead to more consistent, easier to
read specs. There are a few rules that I've relaxed a bit for
convenience.

The only significant departure from the defaults is for request/system
specs. Our style of writing system specs is to mimic how a user would
interact with the app, which usually involves taking several steps:
clicking links, filling out form fields, etc. This isn't really
compatible with RuboCop's preference for extremely short examples.